Added SplitProcess model and UID mixins.
Auto-populates uid from primary key. Fixed several bugs regarding usage
of the tree due to issues with django-mptt. ProcessNodes should always
set parent_id rather than parent to prevent a corruption of the tree
that requires a rebuild as per issue #248. Additionally, whenever the
tree is modified the root element needs to be set on the object instance
to force django to query for the updated tree information.