Added initial implementation of polymorphic serialization.
Working towards issue #101. Added REST fields PolymorphicDataField and
PolymorphicTypeField which are used to represent polymorphic data.

PolymorphicTypeField serializes to the name of the class, useful for
helping the API consumer to know what polymorphic data will be
available.

PolymorphicDataField serializes the polymorphic data itself. It crawls
up the inheritance chain and pulls all of the polymorphic data off into
a flat representation. Currently uses the default serialization for all
types. May break if there are advanced types such as foreignkeys in the
inheritance chain.