Added history to Process and Sample models
Currently created, modified, and user are all unecessary since all of
that information (and more) and be obtained via the history model. The
issue is that these are nontrivial to query on (i.e. find all processes
that were created by a certain user, or list processes in the order of
last modified) so we decided to keep these fields on the model for the
moment.